% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_generations.R
\name{animate_generations}
\alias{animate_generations}
\title{Animate Generational Movement Over a Raster Layer}
\usage{
animate_generations(raster_layer, point_data, output_file)
}
\arguments{
\item{raster_layer}{A Raster* object (RasterLayer, RasterStack, RasterBrick) containing the raster values
to be plotted as the background layer in the animation.}

\item{point_data}{A data frame containing the coordinates and generations for the points. Must contain
the columns 'long', 'lat', and 'gen', where 'gen' is used for the animation frames.}

\item{output_file}{A character string indicating the path and file name for the saved animation file.}
}
\value{
Invisibly returns the animation object created by gganimate::animate().
        The animation is also saved as a file specified by `output_file`.
}
\description{
This function creates an animation of point data moving over a raster layer through time, representing
different generations. The raster is visualized with a 'viridis' color scale and points are animated
over it to show generational changes.
}
