% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_iterate.R
\name{generation_iterate}
\alias{generation_iterate}
\title{Virus Generation Iteration Function}
\usage{
generation_iterate(random_points, mean_dis, R0, input_raster)
}
\arguments{
\item{random_points}{A data frame containing points representing individual virus
particles with latitude, longitude, unique identifiers, and strains.}

\item{mean_dis}{The mean dispersal distance that new points should be generated at,
drawn from a normal distribution.}

\item{R0}{The basic reproduction number per particle, controlling the expected number
of new virus particles generated.}

\item{input_raster}{A raster layer providing spatial context for the simulation,
used to remove generated points that fall onto non-eligible cells like water.}
}
\value{
A data frame of all newly generated points for this iteration, excluding those
        generated over non-eligible areas, with latitude and longitude coordinates,
        identifiers, and associated virus strains.
}
\description{
This function takes a set of points, each representing a virus particle, and simulates
the spread from each of these points according to a Poisson distribution defined by
the basic reproduction number R0. It also accounts for mean dispersal distance to
determine new point locations and ensures that generated points over non-eligible areas
(e.g., water) are removed.
}
