% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_generation.R
\name{initial_generation}
\alias{initial_generation}
\title{Generate Initial Points within a Raster Layer}
\usage{
initial_generation(input_r, n, strain_n)
}
\arguments{
\item{n}{An integer representing the number of random points to generate.}

\item{strain_n}{An integer representing the number of different strain designations to assign.}

\item{test_r}{A raster object representing the spatial area and suitability score for point generation.}
}
\value{
A dataframe of random points with their coordinates, a unique ID, and a strain designation.
}
\description{
This function generates a specified number of random points within the extent of a given raster layer.
The raster values are scaled between 0 and 100, with random points geolocated based on cell values.
Each point is assigned a unique identifier and a random strain designation from a user-defined range.
}
