% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generations.R
\name{plot_generations}
\alias{plot_generations}
\title{Plot Raster Data with Optional Generational Point Overlays}
\usage{
plot_generations(raster_data, point_data = NULL, plot.title = "My Plot Title")
}
\arguments{
\item{raster_data}{A Raster* object (RasterLayer, RasterStack, RasterBrick) containing the raster values.}

\item{point_data}{An optional data frame with columns 'long', 'lat', and 'gen' representing longitude, latitude,
and generational category respectively. Points will be overlaid on the raster and faceted by 'gen'.}

\item{plot.title}{A character string to be used as the main title of the plot.}
}
\value{
A ggplot object representing the raster visualization potentially with generational point overlays.
}
\description{
This function takes raster data to plot it and can optionally overlay point data categorized by generations.
The raster data is represented using a 'viridis' color scale, and the points are plotted in red if provided,
arranged into panels by generations.
}
