% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mappoints.R
\name{plot_mappoints}
\alias{plot_mappoints}
\title{Plot Raster and Optional Point Data}
\usage{
plot_mappoints(raster_data, point_data = NULL, plot.title = "My Plot Title")
}
\arguments{
\item{raster_data}{A Raster* object (RasterLayer, RasterStack, RasterBrick), with raster values to be visualized.}

\item{point_data}{An optional data frame containing point coordinates with columns 'long' for longitude and 'lat' for latitude.
If provided, these points will be overlaid on the raster plot.}

\item{plot.title}{A character string defining the main title of the plot.}
}
\value{
A ggplot object representing the visualized raster data and optional point data overlay.
}
\description{
Generates a plot from raster data, optionally overlaying point data on the raster visualization.
The raster data is visualized using a `viridis` color scale and the points are plotted in red if provided.
}
