% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_raster.R
\name{prep_raster}
\alias{prep_raster}
\title{Prepares a Raster Object}
\usage{
prep_raster(input_raster, scale = TRUE)
}
\arguments{
\item{input_raster}{A SpatRaster object from the terra package representing the raster to be processed.}
}
\value{
A SpatRaster object that has been reprojected and rescaled.
}
\description{
This function reprojects a raster to WGS84 (EPSG:4326) and rescales its cell values to a range of 0 to 100.
}
