% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_cull.R
\name{virus_cull}
\alias{virus_cull}
\title{Virus Population Culling Based on Raster Data}
\usage{
virus_cull(input_r, points_df)
}
\arguments{
\item{input_r}{A RasterLayer object containing values used to determine the probability
of virus particle extinction. High raster values indicate a higher chance of
surviving whereas lower values indicate higher chances of extinction.}

\item{points_df}{A data frame of virus particles with columns for longitude ('long') and
latitude ('lat') representing their geographic locations.}
}
\value{
A data frame containing only the points that have survived (not culled),
        with the 'decisions' column removed.
}
\description{
This function performs culling of a virus population based on raster data values
that represent probabilities of extinction. Given a set of points and associated raster,
it extracts the raster values at each point's location, converts these into probabilities
of survival, and then determines whether each virus particle survives.
}
